+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Support for scatter plots
 * Support for timeline
 * Support for geo charts
+* Support for column charts
 
 ### v0.1.0